mt-cannon: Move thread definitions to new thread.go file
